refactor: New module `type_system`
