+++ lib/skc_ast2hir/src/convert_exprs.rs
@@ -5,7 +5,7 @@ use crate::hir_maker::extract_lvars;
 use crate::hir_maker::HirMaker;
 use crate::hir_maker_context::*;
 use crate::pattern_match;
-use crate::type_checking;
+use crate::type_system::type_checking;
 use anyhow::Result;
 use shiika_ast::Token;
 use shiika_ast::*;
