+++ lib/skc_ast2hir/src/type_system.rs
@@ -1,0 +1,1 @@
+pub mod type_checking;
